requirements.txt
playwright>=1.48.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
python-dotenv>=1.0.0
